Convert remaining breadcrumbs to UBC CLF.
UBC CLF uses Bootstrap's breadcrumb system. The breadcrumb divs needed to be
converted to ul, the id changed to class, and the class to breadcrumb (original
has s at end: breadcrumbs). One thing to note is that the default CAT css
overrides some of the bootstrap breadcrumb cass while instead the div with the
id="content". The workaround is to just move it outside into div
class="wrapper". Note that there was one case where we couldn't do this due to
the breadcrumb being in the middle of a large Java code block. I didn't want to
risk touching those just yet, so the original styling remains.

Also tried to make the footer less noticeable to make it more 'footer-y', as it
felt like it was drawing too much attention away from the main content areas,
especially on shorter pages.